backend/optimizacionRutas/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
google-generativeai==0.4.1
supabase==2.4.0
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
httpx==0.27.0